adding docs; split new file; optimizing on t.loop_get(i)
